tourist_project/requirements.txt
asgiref==3.7.2
autopep8==2.3.2
crispy-bootstrap5==2024.10
Django==5.0.11
django-crispy-forms==2.3
djangorestframework==3.15.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.9.0
